+++ .github/workflows/pythonpublish.yml
@@ -219,7 +219,7 @@ jobs:
           tags: ${{ steps.flyteagent-slim-names.outputs.tags }}
           build-args: |
             VERSION=${{ needs.deploy.outputs.version }}
-          file: ./Dockerfile.agent
+          file: ./Dockerfile.connector
           cache-from: type=gha
           cache-to: type=gha,mode=max
       - name: Confirm Agent can start
@@ -235,7 +235,7 @@ jobs:
           tags: ${{ steps.flyteagent-names.outputs.tags }}
           build-args: |
             VERSION=${{ needs.deploy.outputs.version }}
-          file: ./Dockerfile.agent
+          file: ./Dockerfile.connector
           cache-from: type=gha
           cache-to: type=gha,mode=max
